organized view functions and test functions into separate files (organization) also changed get to queryparams
